requirements.txt
Django==5.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
scikit-learn ==1.5.2
drf-spectacular== 0.9.2